+++ lib/bloom_trade_client.rb
@@ -3,10 +3,15 @@ require 'api_client_base'
 require 'light-service'
 require "message_bus_client_worker"
 require 'virtus'
+
 require 'bloom_trade_client/engine'
 require 'bloom_trade_client/client'
+require 'bloom_trade_client/requests/base_authenticated_request'
 require 'bloom_trade_client/requests/get_quote_request'
+require 'bloom_trade_client/requests/update_quote_request'
+require 'bloom_trade_client/responses/base_response'
 require 'bloom_trade_client/responses/get_quote_response'
+require 'bloom_trade_client/responses/update_quote_response'
 
 module BloomTradeClient