+++ lib/bloom_trade_client/responses/update_quote_response.rb
@@ -1,0 +1,4 @@
+module BloomTradeClient
+  class UpdateQuoteResponse < BaseResponse
+  end
+end