+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic
 Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Add `BloomTradeClient::Client#update_quote`
+- Refactor rate type calculations on `BloomTradeClient::ExchangeRates::Convert`
+
 ## [0.11.0] - 2018-08-16
 ### Changed
 - Renamed gem into `bloom_trade_client-rails` from `bloom_rates-rails`